+++ @xen-orchestra/web/src/modules/pool/components/list/panel/PoolSidePanel.vue
@@ -1,5 +1,5 @@
 <template>
-  <VtsSidePanel :has-selection="!!server" @close="emit('close')">
+  <VtsSidePanel :has-selection="!!server" class="pool-side-panel" @close="emit('close')">
     <template v-if="server" #actions>
       <PoolConnectionToggleButton :server-id="server.id" />
       <MenuList placement="bottom-end">
@@ -19,7 +19,7 @@
     <template v-if="server">
       <VtsStateHero v-if="!arePoolsReady" format="panel" type="busy" size="medium" />
       <template v-else>
-        <UiCard v-if="server.error === undefined" class="card-container">
+        <UiPanelCard v-if="server.error === undefined">
           <VtsCardObjectTitle :id="server.id" :label="server.label" icon="object:pool" />
           <div class="content">
             <!-- Pool -->
@@ -60,14 +60,14 @@
               </template>
             </VtsCardRowKeyValue>
           </div>
-        </UiCard>
+        </UiPanelCard>
         <UiAlert v-else accent="danger">
           {{ t('connection-failed') }}
           <template #description>
             {{ t('unable-to-connect-to-the-pool') }}
           </template>
         </UiAlert>
-        <UiCard class="card-container">
+        <UiPanelCard>
           <UiCardTitle>
             {{ t('connection') }}
           </UiCardTitle>
@@ -138,8 +138,8 @@
               <VtsStatus :status="server.allowUnauthorized" />
             </template>
           </VtsCardRowKeyValue>
-        </UiCard>
-        <UiCard v-if="hosts !== undefined">
+        </UiPanelCard>
+        <UiPanelCard v-if="hosts !== undefined">
           <UiCardTitle>
             <span>
               {{ t('hosts') }}
@@ -161,14 +161,14 @@
               <VtsIcon v-if="primaryHost?.id === host.id" accent="info" name="status:primary-circle" size="medium" />
             </UiLink>
           </template>
-        </UiCard>
-        <UiCard v-if="server.error">
+        </UiPanelCard>
+        <UiPanelCard v-if="server.error">
           <UiCardTitle>
             {{ t('error') }}
             <UiCounter :value="1" accent="danger" size="small" variant="primary" />
           </UiCardTitle>
           <UiLogEntryViewer accent="danger" :label="t('api-error-details')" size="small" :content="server.error" />
-        </UiCard>
+        </UiPanelCard>
       </template>
     </template>
   </VtsSidePanel>
@@ -192,12 +192,12 @@ import VtsStatus from '@core/components/status/VtsStatus.vue'
 import VtsTag from '@core/components/tag/VtsTag.vue'
 import UiAlert from '@core/components/ui/alert/UiAlert.vue'
 import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
-import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiCardTitle from '@core/components/ui/card-title/UiCardTitle.vue'
 import UiCounter from '@core/components/ui/counter/UiCounter.vue'
 import UiInfo from '@core/components/ui/info/UiInfo.vue'
 import UiLink from '@core/components/ui/link/UiLink.vue'
 import UiLogEntryViewer from '@core/components/ui/log-entry-viewer/UiLogEntryViewer.vue'
+import UiPanelCard from '@core/components/ui/panel-card/UiPanelCard.vue'
 import UiTagsList from '@core/components/ui/tag/UiTagsList.vue'
 import { vTooltip } from '@core/directives/tooltip.directive.ts'
 import { useMapper } from '@core/packages/mapper'
@@ -234,11 +234,7 @@ const connectionStatus = useMapper(
 </script>
 
 <style scoped lang="postcss">
-.card-container {
-  display: flex;
-  flex-direction: column;
-  gap: 1.6rem;
-
+.pool-side-panel {
   .content {
     display: flex;
     flex-direction: column;